requirements.txt
numpy>=1.26,<3
pandas>=2.0,<3
tqdm>=4.66,<5
sentence-transformers>=2.7,<3
faiss-cpu>=1.7,<2
streamlit>=1.36,<2
python-dotenv>=1.0,<2
fastapi>=0.110,<1  # existing FastAPI app (optional)
uvicorn>=0.30,<1
